requirements.txt
flask==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
flask-login==0.6.3
flask-wtf==1.2.1
werkzeug==3.0.1
openai-whisper==20231117
torch==2.2.1
transformers==4.38.2
xlsxwriter==3.1.9 